Remove unnecessary nots in schema
"not": { "required": ["value", "valueStart"] }
was used to forbid mixing step and proportional value.

However, when the value is required for the capability, there's a
oneOf that ensures that either the step or proportional value
is used:

"oneOf": [
  { "required": ["value"] },
  { "required": ["valueStart"] }
],

This already includes that not both of them can be true
and makes the not property obsolete.

This commit removes these obsolete not properties.