Refactor dmxRange in model, add rawDmxRange
The helper method getDmxRangeWithFineness doesn't rely on the (upscaled)
property dmxRange anymore, but on the raw JSON dmxRange. It also caches
values now.

dmxRange becomes only a shorthand for getDmxRangeWithFineness with the
channel's maximum fineness as parameter.

rawDmxRange is a shorthand for getDmxRangeWithFineness with the fineness
used in the raw JSON as parameter. It equals the JSON's dmxRange.