Flip Chroma-Q Color Force II's matrix direction
There is no clear mention in the manual that the "highest" pixel
should not be on the right. Actually, the table in the manual
is left to right, so also using that direction makes sense.

Additionally, the direction can be flipped in the device menu
anyway, so we should focus more on readability.