Fix SVG loader on Windows (#1276)
For proper packing the SVGs in ui/assets/icons must not use the
standard url-loader and instead use the svg-inline-loader.
Previously this was achieved by an exclude for url-loader
containing a RegExp with a forward slash. However Windows uses
backwards slashes for paths, causing the RegExp to not work.

This commit changed the exclude to use the Node path module,
as suggested in https://webpack.js.org/configuration/#options