Add Windows Install Instructions (#1276)
Closes #1276.

I also reworked the Linux instructions to include the dependency on
git, make and Node. Also converted the instructions to
copy-pastable commands.
Of course, one might clone into a differently named folder, but when
someone knows how to do this I would expect them to also adapth the cd
command without a problem.