Use plain "2" for "H2O"
Eurolite refers to this product in various places as "H²O" [sic], "H₂O", and (most commonly) "H2O".  I propose that, in the absence of a strong need to use a superscript (or subscript), we simply call it by the plain "H2O", as it's easiest to type, and to search for.