Restore site-crawler's `stopServer` function
In Node 18+ (maybe because of changes in npm or npx), the server
process could not be terminated with `kill` alone. Apparently, a sub-
process is launched, which is not stopped when the direct child
process is killed. It seems like installing the `terminate` dependency
is the easiest option to fix this.