Update rogue-r1-wash beam angle
Update the zoom to use beam angle, as well

So far, the "lens" had 8...30 degrees, and the "zoom" had 11...48 degrees.  According to the specs, 8-30 is the beam angle, and 11-48 is the field angle (and also the "zoom angle", whatever that means).

In schemas/capability.json:816, zoom refers to "beamAngle".  In lib/model/Physical.js:120-132, degreesMinMax refers to beam angle, as well.