Update uking/b117 switched channel
Previously, the switch at value 0 selected "Color Presets", which was a channel entirely populated with ColorPreset capabilities.

My understanding is that setting Effect=0=NoFunction would thus appear (to any application/user relying on this mapping) to enable the ColorPresets, which isn't how this fixture works.

This updates the switch to point to a dead channel when a non-effect is chosen.  I've found precedent for this in OFL, e.g., `varytec/giga-bar-frost-pix-8-rgb`.